Refactor acceptance tests splitting to respect coding style
[MAILPOET-3296]
